This reverts commit dcc471b684764731f86c58efe4a67e9e4c1bafae.

+++ Heinrich/src/view/InputDialog.java
@@ -220,7 +220,7 @@ public class InputDialog extends JDialog
 	{
 		try
 		{
-			Float.parseFloat(input);
+			Double.parseDouble(input);
 			return true;
 		} catch (NumberFormatException e)
 		{
